ログインしているユーザーのみ削除できる機能

+++ app/controllers/items_controller.rb
@@ -1,5 +1,5 @@
 class ItemsController < ApplicationController
-  before_action :authenticate_user!, only: [:new, :edit]
+  before_action :authenticate_user!, only: [:new, :edit, :destroy]
   before_action :set_item, only: [:edit, :show, :update, :destroy]
 
   def index
@@ -38,13 +38,14 @@ class ItemsController < ApplicationController
   end
 
   def destroy
-
+if user_signed_in?
     if @item.destroy
       redirect_to root_path
 
     else
       render :show
     end
+   end
   end
 
   private
